requirements.txt
streamlit
torch
torchvision
timm
segmentation-models-pytorch
numpy
opencv-python-headless
Pillow
pandas
cloudinary
gspread
google-auth
reportlab
requests